+++ packages/plugin-collapsible-list/web/src/components/modal-components/CollapsibleListSettings.jsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { mergeStyles } from 'wix-rich-content-common';
-import { Separator } from 'wix-rich-content-editor-common';
 import {
   RadioGroupVertical,
   SelectionList,
+  Separator,
   LabeledToggle,
   InfoIcon,
 } from 'wix-rich-content-ui-components';