+++ packages/common/web/src/types/biCallbacksTypes.ts
@@ -44,7 +44,7 @@ export interface onViewerLoadedArgs extends biCallbackParams {
   isPreview: boolean;
   pluginsCount: ReturnType<typeof getContentSummary>['pluginsCount'];
   version: string;
-  postURL: string;
+  url: string;
 }
 
 export interface onPluginModalOpenedArgs extends biCallbackParams {